+++ spec/factories/assets.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :asset do
+    filename { fixture_file_upload 'spec/fixtures/files/example.jpg', 'application/jpeg' }
+  end
+end